move mock_transport.go into peer_test.go
